Update index - lghales.qmd
The following is an assessment of the code as of 11/19/2024

The code provided for the RCP 8.5 and 2.6 analysis is very extensive and detailed with a great clean up and provides a smaller but still significantly detailed methods and results section of the ice sheet model. The only complaint I have is with the formatting of the introduction and Materials & Methods section of the code, as the text is extended across the webpage. Besides that, the code is a great analysis.